Extract base class to make adding new message types easier
